MODELTEST/docker_shared/requirements.txt
pandas
scikit-learn
xgboost
joblib
Flask
requests
google-cloud-storage

